moving the flex reparent code to a shared function
