+++ editor/src/components/canvas/canvas-strategies/absolute-reparent-to-flex-strategy.tsx
@@ -27,7 +27,11 @@ import { updateHighlightedViews } from '../commands/update-highlighted-views-com
 import { CanvasCommand, foldAndApplyCommandsInner } from '../commands/commands'
 import { deleteProperties } from '../commands/delete-properties-command'
 import { updateSelectedViews } from '../commands/update-selected-views-command'
-import { findReparentStrategy, getReparentTargetForFlexElement } from './reparent-strategy-helpers'
+import {
+  applyFlexReparent,
+  findReparentStrategy,
+  getReparentTargetForFlexElement,
+} from './reparent-strategy-helpers'
 
 const propertiesToRemove: Array<PropertyPath> = [
   PP.create(['style', 'position']),
@@ -102,74 +106,6 @@ export const absoluteReparentToFlexStrategy: CanvasStrategy = {
     interactionSession: InteractionSession,
     strategyState: StrategyState,
   ): StrategyApplicationResult {
-    if (
-      interactionSession.interactionData.type == 'DRAG' &&
-      interactionSession.interactionData.drag != null
-    ) {
-      const filteredSelectedElements = getDragTargets(canvasState.selectedElements)
-      const reparentResult = getReparentTargetForFlexElement(
-        filteredSelectedElements,
-        interactionSession,
-        canvasState,
-        strategyState,
-      )
-
-      if (
-        reparentResult.shouldReparent &&
-        reparentResult.newParent != null &&
-        filteredSelectedElements.length === 1
-      ) {
-        const target = filteredSelectedElements[0]
-        const newParent = reparentResult.newParent
-        // Reparent the element.
-        const newPath = EP.appendToPath(reparentResult.newParent, EP.toUid(target))
-        const reparentCommand = reparentElement('permanent', target, reparentResult.newParent)
-
-        // Strip the `position`, positional and dimension properties.
-        const commandToRemoveProperties = deleteProperties('permanent', newPath, propertiesToRemove)
-
-        const commandsBeforeReorder = [reparentCommand, updateSelectedViews('permanent', [newPath])]
-
-        const commandsAfterReorder = [
-          commandToRemoveProperties,
-          setElementsToRerenderCommand([newPath]),
-          updateHighlightedViews('transient', []),
-          setCursorCommand('transient', CSSCursor.Move),
-        ]
-
-        let commands: Array<CanvasCommand>
-        if (reparentResult.shouldReorder) {
-          // Reorder the newly reparented element into the flex ordering.
-          const pointOnCanvas = offsetPoint(
-            interactionSession.interactionData.dragStart,
-            interactionSession.interactionData.drag,
-          )
-
-          const siblingsOfTarget = MetadataUtils.getChildrenPaths(
-            strategyState.startingMetadata,
-            newParent,
-          )
-
-          const newIndex = getReorderIndex(
-            strategyState.startingMetadata,
-            siblingsOfTarget,
-            pointOnCanvas,
-          )
-          commands = [
-            ...commandsBeforeReorder,
-            reorderElement('permanent', newPath, newIndex),
-            ...commandsAfterReorder,
-          ]
-        } else {
-          commands = [...commandsBeforeReorder, ...commandsAfterReorder]
-        }
-
-        return {
-          commands: commands,
-          customState: strategyState.customStrategyState,
-        }
-      }
-    }
-    return emptyStrategyApplicationResult
+    return applyFlexReparent('strip-absolute-props', canvasState, interactionSession, strategyState)
   },
 }
