+++ editor/src/components/canvas/canvas-strategies/flex-reparent-to-flex-strategy.tsx
@@ -14,7 +14,11 @@ import { ParentOutlines } from '../controls/parent-outlines'
 import { DragOutlineControl } from '../controls/select-mode/drag-outline-control'
 import { CanvasStrategy, emptyStrategyApplicationResult } from './canvas-strategy-types'
 import { getReorderIndex } from './flex-reorder-strategy'
-import { findReparentStrategy, getReparentTargetForFlexElement } from './reparent-strategy-helpers'
+import {
+  applyFlexReparent,
+  findReparentStrategy,
+  getReparentTargetForFlexElement,
+} from './reparent-strategy-helpers'
 import { getDragTargets } from './shared-absolute-move-strategy-helpers'
 
 export const flexReparentToFlexStrategy: CanvasStrategy = {
@@ -60,70 +64,6 @@ export const flexReparentToFlexStrategy: CanvasStrategy = {
     return 0
   },
   apply: (canvasState, interactionSession, strategyState) => {
-    if (
-      interactionSession.interactionData.type == 'DRAG' &&
-      interactionSession.interactionData.drag != null
-    ) {
-      const filteredSelectedElements = getDragTargets(canvasState.selectedElements)
-      const reparentResult = getReparentTargetForFlexElement(
-        filteredSelectedElements,
-        interactionSession,
-        canvasState,
-        strategyState,
-      )
-
-      if (
-        reparentResult.shouldReparent &&
-        reparentResult.newParent != null &&
-        filteredSelectedElements.length === 1
-      ) {
-        const target = filteredSelectedElements[0]
-        const newParent = reparentResult.newParent
-        // Reparent the element.
-        const newPath = EP.appendToPath(reparentResult.newParent, EP.toUid(target))
-        const reparentCommand = reparentElement('permanent', target, reparentResult.newParent)
-
-        const commandsBeforeReorder = [reparentCommand, updateSelectedViews('permanent', [newPath])]
-
-        const commandsAfterReorder = [
-          setElementsToRerenderCommand([newPath]),
-          updateHighlightedViews('transient', []),
-          setCursorCommand('transient', CSSCursor.Move),
-        ]
-
-        let commands: Array<CanvasCommand>
-        if (reparentResult.shouldReorder) {
-          // Reorder the newly reparented element into the flex ordering.
-          const pointOnCanvas = offsetPoint(
-            interactionSession.interactionData.dragStart,
-            interactionSession.interactionData.drag,
-          )
-
-          const siblingsOfTarget = MetadataUtils.getChildrenPaths(
-            strategyState.startingMetadata,
-            newParent,
-          )
-
-          const newIndex = getReorderIndex(
-            strategyState.startingMetadata,
-            siblingsOfTarget,
-            pointOnCanvas,
-          )
-          commands = [
-            ...commandsBeforeReorder,
-            reorderElement('permanent', newPath, newIndex),
-            ...commandsAfterReorder,
-          ]
-        } else {
-          commands = [...commandsBeforeReorder, ...commandsAfterReorder]
-        }
-
-        return {
-          commands: commands,
-          customState: strategyState.customStrategyState,
-        }
-      }
-    }
-    return emptyStrategyApplicationResult
+    return applyFlexReparent('do-not-strip-props', canvasState, interactionSession, strategyState)
   },
 }
